Pre requirements to use this vscode extension
Hi,

I faced a question mark when I start to use this extension because I suppose that all dependencies was setup. However this project only wrap the yo CLI inside the vscode. So I suggest this change to include a topic to explain the pre-requirements to use it to new users which are not familiar with this ecosystem of tools.

Thank you,
Rafael